pyproject.toml
[project]
name = "paperbreakfast"
version = "0.1.0"
description = "Scientific paper curation via RSS + LLM"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.0",
    "peewee>=3.17.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
paperbreakfast = "paperbreakfast.cli:main"

[tool.setuptools.packages.find]
include = ["paperbreakfast*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
